implement partial_gui wip
