+++ magicclass/_gui/toolbar.py
@@ -15,11 +15,10 @@ from ._base import (
     PopUpMode,
     ErrorMode,
     ContainerLikeGui,
-    nested_function_gui_callback,
-    _inject_recorder,
 )
 from .utils import format_error, connect_magicclasses
 from .menu_gui import ContextMenuGui, MenuGui, MenuGuiBase, insert_action_like
+from ._macro_utils import inject_recorder, nested_function_gui_callback
 
 from ..signature import get_additional_option
 from ..fields import MagicField
@@ -237,7 +236,7 @@ class ToolBarGui(ContainerLikeGui):
                     obj.called.connect(f)
                 _obj = obj
             else:
-                obj = _inject_recorder(obj, is_method=False).__get__(self)
+                obj = inject_recorder(obj, is_method=False).__get__(self)
                 _obj = self._create_widget_from_method(obj)
 
             method_name = getattr(obj, "__name__", None)
