+++ magicclass/utils/__init__.py
@@ -22,5 +22,6 @@ from .qt import (
 )
 
 from ._click import click
+from ._partial import partial_gui
 from .qtsignal import QtSignal
 from .qthreading import thread_worker, Timer, Callback
